Methods Override
Method Overriding in Java: 
If the child class implements the same method present in the parent class again, it is know as method overriding.
Method overriding helps us to classify a behavior that is specific to the child class.
The subclass can override the method of the parent class only when the method is not declared as final. 
Example :
In the below code, we've created two classes: class A & class B.
Class B is inheriting class A.
In the main() method, we've created one object for both classes. We're running the meth1() method on class A and B objects separately, but the output is the same because the meth1() is defined in the parent class, i.e., class A.