Add files via upload
Java Generics
Introduced from JDK 5.0 onwards.
The Java Generics helps us to deal with the compiler time type-safety.
With the help of the Generics, we can write a single method and call it with different argument types(integer, strings, etc.).
Advantages of Generics :
Bugs can be detected at compile-time:
While developing any application or program, it is always better to catch the bug/problem at the compile-time instead of runtime so that we can provide a smooth experience to the user.
Let's take an example to see how Java Generics helps us to detect problems at compile-time:
Example :
import java.util.ArrayList;

public class CWH {
    public static void main(String[] args) {
//        Without Java Generics :

        ArrayList myArrayList = new ArrayList();
        myArrayList.add(10);  //Integer value
        myArrayList.add("Harry Bhai!"); //String value
        myArrayList.add(20.4); //Double value
        System.out.println(myArrayList);

    }
}
Output :
[10, Harry Bhai!, 20.4]
In the above code, notice that we can store any type of object in a collection without Generics. But, this is not the case with the Generics. It allows us to store only one type of object. Take a look at the example given below :

import java.util.ArrayList;

public class CWH {
    public static void main(String[] args) {
//        With Java Generics :

        ArrayList<Integer> myArrayList = new ArrayList();
        myArrayList.add(10);  //Integer value
        myArrayList.add("Harry Bhai!"); //String value
        myArrayList.add(20.4); //Double value
        System.out.println(myArrayList);

    }
}
Output :

java: incompatible types: java.lang.String cannot be converted to java.lang.Integer
java: incompatible types: double cannot be converted to java.lang.Integer
The same code produces the Incompatible type error because we can only store the integer object type.

Type-casting not required :
Let's suppose you created an ArrayList(without using Generics), and you want to store the value at index 0 into an integer variable named "x." Are you allowed to do this in Java? The answer is a big NO! This is because the ArrayList returns an object, but we're storing the value in an integer variable. In such cases, we need to type-cast the object into our desired data type. But, if we use Generics, then there is no need to typecast. Take a look at the below example to get a better understanding :
import java.util.ArrayList;

public class CWH {
    public static void main(String[] args) {
//        Without Java Generics :

        ArrayList myArrayList = new ArrayList();
        myArrayList.add(10);  //Integer value
        myArrayList.add("Harry Bhai!"); //String value
        myArrayList.add(20.4); //Double value

        int x = myArrayList.get(0);
        System.out.println(x);

    }
}
Output :
java: incompatible types: java.lang.Object cannot be converted to int
The above code produces an error because we've not typecasted the object into the integer type. Now, let's typecast and see the results :

import java.util.ArrayList;

public class CWH {
    public static void main(String[] args) {
//        Without Java Generics :

        ArrayList myArrayList = new ArrayList();
        myArrayList.add(10);  //Integer value
        myArrayList.add("Harry Bhai!"); //String value
        myArrayList.add(20.4); //Double value

        int x = (int) myArrayList.get(0); //b=object typecasted into integer
        System.out.println(x);

    }

}
Output :
10
Now, let's see how we can get the desired results with the help of the Generics :

import java.util.ArrayList;

public class CWH {
    public static void main(String[] args) {
//        With Java Generics :

        ArrayList<Integer> myArrayList = new ArrayList();
        myArrayList.add(10);
        myArrayList.add(20);
        myArrayList.add(30);
        myArrayList.add(40);

        int x = myArrayList.get(0);
        System.out.println(x);

    }
}
Output :
10